rename Guppy module and update references to Normalize
