+++ tket/src/passes.rs
@@ -33,10 +33,10 @@ pub use dead_funcs::{RemoveDeadFuncsError, RemoveDeadFuncsPass};
 pub mod force_order;
 
 // Normalize the structure of programs.
-pub mod guppy;
-pub use guppy::Normalize;
+pub mod normalize;
+pub use normalize::Normalize;
 #[expect(deprecated)]
-pub use guppy::NormalizeGuppy;
+pub use normalize::NormalizeGuppy;
 
 // Inline DFG nodes.
 pub mod inline_dfgs;
