+++ tket/src/passes/normalize.rs
